MonitorSpecificService/requirements.txt
google-cloud-logging==2.4.0
pg8000==1.23.0
SQLAlchemy==1.4.29
uuid==1.30